+++ client-runtime/src/test/java/com/microsoft/rest/v2/http/HttpPipelineBuilderTests.java
@@ -1,7 +1,7 @@
 package com.microsoft.rest.v2.http;
 
 import com.microsoft.rest.v2.policy.LoggingPolicy;
-import com.microsoft.rest.v2.policy.PortPolicy;
+import com.microsoft.rest.v2.policy.PortPolicyFactory;
 import com.microsoft.rest.v2.policy.ProtocolPolicyFactory;
 import com.microsoft.rest.v2.policy.RetryPolicyFactory;
 import org.junit.Test;
@@ -21,38 +21,38 @@ public class HttpPipelineBuilderTests {
     public void withRequestPolicy() {
         final HttpPipelineBuilder builder = new HttpPipelineBuilder();
 
-        builder.withRequestPolicy(new PortPolicy.Factory(80));
+        builder.withRequestPolicy(new PortPolicyFactory(80));
         assertEquals(1, builder.requestPolicyFactories().size());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
+        assertEquals(PortPolicyFactory.class, builder.requestPolicyFactories().get(0).getClass());
 
         builder.withRequestPolicy(new ProtocolPolicyFactory("ftp"));
         assertEquals(2, builder.requestPolicyFactories().size());
         assertEquals(ProtocolPolicyFactory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
+        assertEquals(PortPolicyFactory.class, builder.requestPolicyFactories().get(1).getClass());
 
         builder.withRequestPolicy(new RetryPolicyFactory());
         assertEquals(3, builder.requestPolicyFactories().size());
         assertEquals(RetryPolicyFactory.class, builder.requestPolicyFactories().get(0).getClass());
         assertEquals(ProtocolPolicyFactory.class, builder.requestPolicyFactories().get(1).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(2).getClass());
+        assertEquals(PortPolicyFactory.class, builder.requestPolicyFactories().get(2).getClass());
     }
 
     @Test
     public void withRequestPolicyWithIndex() {
         final HttpPipelineBuilder builder = new HttpPipelineBuilder();
 
-        builder.withRequestPolicy(0, new PortPolicy.Factory(80));
+        builder.withRequestPolicy(0, new PortPolicyFactory(80));
         assertEquals(1, builder.requestPolicyFactories().size());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
+        assertEquals(PortPolicyFactory.class, builder.requestPolicyFactories().get(0).getClass());
 
         builder.withRequestPolicy(0, new ProtocolPolicyFactory("ftp"));
         assertEquals(2, builder.requestPolicyFactories().size());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
+        assertEquals(PortPolicyFactory.class, builder.requestPolicyFactories().get(0).getClass());
         assertEquals(ProtocolPolicyFactory.class, builder.requestPolicyFactories().get(1).getClass());
 
         builder.withRequestPolicy(1, new RetryPolicyFactory());
         assertEquals(3, builder.requestPolicyFactories().size());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
+        assertEquals(PortPolicyFactory.class, builder.requestPolicyFactories().get(0).getClass());
         assertEquals(RetryPolicyFactory.class, builder.requestPolicyFactories().get(1).getClass());
         assertEquals(ProtocolPolicyFactory.class, builder.requestPolicyFactories().get(2).getClass());
     }
@@ -63,12 +63,12 @@ public class HttpPipelineBuilderTests {
 
         builder.withRequestPolicies(
                 new ProtocolPolicyFactory("http"),
-                new PortPolicy.Factory(80),
+                new PortPolicyFactory(80),
                 new LoggingPolicy.Factory(LoggingPolicy.LogLevel.BODY));
 
         assertEquals(3, builder.requestPolicyFactories().size());
         assertEquals(LoggingPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
+        assertEquals(PortPolicyFactory.class, builder.requestPolicyFactories().get(1).getClass());
         assertEquals(ProtocolPolicyFactory.class, builder.requestPolicyFactories().get(2).getClass());
     }
 
@@ -79,12 +79,12 @@ public class HttpPipelineBuilderTests {
         builder.withRequestPolicy(new RetryPolicyFactory());
         builder.withRequestPolicies(
                 new ProtocolPolicyFactory("http"),
-                new PortPolicy.Factory(80),
+                new PortPolicyFactory(80),
                 new LoggingPolicy.Factory(LoggingPolicy.LogLevel.BODY));
 
         assertEquals(4, builder.requestPolicyFactories().size());
         assertEquals(LoggingPolicy.Factory.class, builder.requestPolicyFactories().get(0).getClass());
-        assertEquals(PortPolicy.Factory.class, builder.requestPolicyFactories().get(1).getClass());
+        assertEquals(PortPolicyFactory.class, builder.requestPolicyFactories().get(1).getClass());
         assertEquals(ProtocolPolicyFactory.class, builder.requestPolicyFactories().get(2).getClass());
         assertEquals(RetryPolicyFactory.class, builder.requestPolicyFactories().get(3).getClass());
